requirements.txt
absl-py==2.3.0
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argbind==0.3.9
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
billiard==4.2.1
boto3==1.38.27
botocore==1.38.27
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cron-descriptor==1.4.5
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==2.12.0
deprecation==2.1.0
descript-audio-codec==1.0.0
distro==1.9.0
Django==5.2.1
django-celery-beat==2.8.1
django-enumfields==2.1.1
django-json-widget==2.0.1
django-ninja==1.4.2
django-timezone-field==7.1
docstring_parser==0.16
einops==0.8.1
executing==2.2.0
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
greenlet==3.2.2
grpcio==1.72.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h3==4.2.2
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
influxdb3-python==0.12.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.25
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langid==1.1.6
langsmith==0.3.43
lazy_loader==0.4
librosa==0.11.0
litellm==1.72.0
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
openai==1.82.1
orjson==3.10.18
packaging==24.2
parler_tts @ git+https://github.com/huggingface/parler-tts.git@d108732cd57788ec86bc857d99a6cabd66663d68
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
pystoi==0.4.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
randomname==0.2.1
reactivex==4.0.4
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
s3transfer==0.13.0
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
smallestai==2.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timezonefinder==6.5.9
tokenizers==0.20.3
torch==2.6.0
torch-stoi==0.2.3
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
